+++ src/main/java/me/william278/huskhomes2/api/events/PlayerDeleteWarpEvent.java
@@ -1,6 +1,6 @@
 package me.william278.huskhomes2.api.events;
 
-import me.william278.huskhomes2.teleport.Warp;
+import me.william278.huskhomes2.teleport.points.Warp;
 import org.bukkit.entity.Player;
 
 /**
