+++ src/main/java/me/william278/huskhomes2/api/events/PlayerRelocateHomeEvent.java
@@ -1,7 +1,7 @@
 package me.william278.huskhomes2.api.events;
 
-import me.william278.huskhomes2.teleport.Home;
-import me.william278.huskhomes2.teleport.TeleportationPoint;
+import me.william278.huskhomes2.teleport.points.Home;
+import me.william278.huskhomes2.teleport.points.TeleportationPoint;
 import org.bukkit.entity.Player;
 
 /**
