+++ src/main/java/me/william278/huskhomes2/commands/BackCommand.java
@@ -1,9 +1,9 @@
 package me.william278.huskhomes2.commands;
 
-import me.william278.huskhomes2.data.DataManager;
 import me.william278.huskhomes2.MessageManager;
-import me.william278.huskhomes2.teleport.TeleportationPoint;
+import me.william278.huskhomes2.data.DataManager;
 import me.william278.huskhomes2.teleport.TeleportManager;
+import me.william278.huskhomes2.teleport.points.TeleportationPoint;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
