+++ src/main/java/me/william278/huskhomes2/commands/HomeCommand.java
@@ -1,10 +1,10 @@
 package me.william278.huskhomes2.commands;
 
-import me.william278.huskhomes2.data.DataManager;
-import me.william278.huskhomes2.ListHandler;
 import me.william278.huskhomes2.MessageManager;
-import me.william278.huskhomes2.teleport.Home;
+import me.william278.huskhomes2.data.DataManager;
+import me.william278.huskhomes2.teleport.ListHandler;
 import me.william278.huskhomes2.teleport.TeleportManager;
+import me.william278.huskhomes2.teleport.points.Home;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
