+++ src/main/java/me/william278/huskhomes2/commands/TpCommand.java
@@ -1,10 +1,10 @@
 package me.william278.huskhomes2.commands;
 
 import me.william278.huskhomes2.HuskHomes;
-import me.william278.huskhomes2.integrations.VanishChecker;
 import me.william278.huskhomes2.MessageManager;
-import me.william278.huskhomes2.teleport.TeleportationPoint;
+import me.william278.huskhomes2.integrations.VanishChecker;
 import me.william278.huskhomes2.teleport.TeleportManager;
+import me.william278.huskhomes2.teleport.points.TeleportationPoint;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
