+++ src/main/java/me/william278/huskhomes2/integrations/DynMapIntegration.java
@@ -1,9 +1,9 @@
 package me.william278.huskhomes2.integrations;
 
 import me.william278.huskhomes2.HuskHomes;
-import me.william278.huskhomes2.teleport.Home;
-import me.william278.huskhomes2.teleport.Warp;
 import me.william278.huskhomes2.data.DataManager;
+import me.william278.huskhomes2.teleport.points.Home;
+import me.william278.huskhomes2.teleport.points.Warp;
 import org.bukkit.plugin.Plugin;
 import org.dynmap.DynmapAPI;
 import org.dynmap.markers.Marker;
