+++ src/main/java/me/william278/huskhomes2/listeners/PlayerListener.java
@@ -1,11 +1,11 @@
 package me.william278.huskhomes2.listeners;
 
 import me.william278.huskhomes2.HuskHomes;
+import me.william278.huskhomes2.MessageManager;
 import me.william278.huskhomes2.commands.HomeCommand;
 import me.william278.huskhomes2.data.DataManager;
-import me.william278.huskhomes2.MessageManager;
-import me.william278.huskhomes2.teleport.TeleportationPoint;
 import me.william278.huskhomes2.teleport.TeleportManager;
+import me.william278.huskhomes2.teleport.points.TeleportationPoint;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
