+++ src/main/java/me/william278/huskhomes2/teleport/ListHandler.java
@@ -1,8 +1,10 @@
-package me.william278.huskhomes2;
+package me.william278.huskhomes2.teleport;
 
+import me.william278.huskhomes2.HuskHomes;
+import me.william278.huskhomes2.MessageManager;
 import me.william278.huskhomes2.data.DataManager;
-import me.william278.huskhomes2.teleport.Home;
-import me.william278.huskhomes2.teleport.Warp;
+import me.william278.huskhomes2.teleport.points.Home;
+import me.william278.huskhomes2.teleport.points.Warp;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
