+++ src/main/java/me/william278/huskhomes2/teleport/points/RandomPoint.java
@@ -1,4 +1,4 @@
-package me.william278.huskhomes2.teleport;
+package me.william278.huskhomes2.teleport.points;
 
 import me.william278.huskhomes2.HuskHomes;
 import me.william278.huskhomes2.MessageManager;
